src/projects/mibs/requirements.txt
flask
python-dotenv
requests
Flask-SQLAlchemy
psycopg2
pyjwt[crypto]